Update the README with my current plans.
I'm still new to Jetpack Compose and some of the other recently-added
Android Jetpack components, so I've decided to write a README.md
for the top level and a more detailed version in src/README.md with
my plans for the classes to write.

Hopefully I can make some progress today and over the weekend on
implementing the game that I can see in my mind's eye.